+++ plugins/destination/postgresql/client/read.go
@@ -9,7 +9,7 @@ import (
 )
 
 const (
-	readSQL = "SELECT * FROM %s WHERE _cq_source_name = $1"
+	readSQL = "SELECT * FROM %s WHERE _cq_source_name = $1 order by _cq_sync_time asc"
 )
 
 func (c *Client) Read(ctx context.Context, table *schema.Table, sourceName string, res chan<- []any) error {