+++ plugins/source/aws/resources/services/rds/instances.go
@@ -273,6 +273,11 @@ func Instances() *schema.Table {
 				Resolver: schema.PathResolver("ListenerEndpoint"),
 			},
 			{
+				Name:     "master_user_secret",
+				Type:     schema.TypeJSON,
+				Resolver: schema.PathResolver("MasterUserSecret"),
+			},
+			{
 				Name:     "master_username",
 				Type:     schema.TypeString,
 				Resolver: schema.PathResolver("MasterUsername"),