+++ plugins/source/gcp/codegen/recipes/serviceusage.go
@@ -8,16 +8,15 @@ import (
 func init() {
 	resources := []*Resource{
 		{
-			SubService:          "services",
-			Struct:              &pb.Service{},
-			NewFunction:         serviceusage.NewClient,
-			RequestStruct:       &pb.ListServicesRequest{},
-			ResponseStruct:      &pb.ListServicesResponse{},
-			RegisterServer:      pb.RegisterServiceUsageServer,
-			ListFunction:        (&pb.UnimplementedServiceUsageServer{}).ListServices,
-			UnimplementedServer: &pb.UnimplementedServiceUsageServer{},
-			PrimaryKeys:         []string{"name"},
-			Description:         "https://cloud.google.com/service-usage/docs/reference/rest/v1/services#Service",
+			SubService:     "services",
+			Struct:         &pb.Service{},
+			NewFunction:    serviceusage.NewClient,
+			RegisterServer: pb.RegisterServiceUsageServer,
+			ListFunction:   (&pb.UnimplementedServiceUsageServer{}).ListServices,
+			PrimaryKeys:    []string{"name"},
+			Description:    "https://cloud.google.com/service-usage/docs/reference/rest/v1/services#Service",
+			SkipFetch:      true,
+			SkipMock:       true,
 		},
 	}
 
@@ -27,9 +26,7 @@ func init() {
 		resource.ProtobufImport = "cloud.google.com/go/serviceusage/apiv1/serviceusagepb"
 		resource.Template = "newapi_list"
 		resource.MockTemplate = "newapi_list_grpc_mock"
-		resource.RequestStructFields = `Parent: "projects/" + c.ProjectId,
-		PageSize: 200,
-		Filter: "state:ENABLED",`
+		resource.ServiceDNS = "serviceusage.googleapis.com"
 	}
 
 	Resources = append(Resources, resources...)