+++ plugins/source/gitlab/client/services/settings.go
@@ -1,0 +1,7 @@
+package services
+
+import "github.com/xanzy/go-gitlab"
+
+type SettingsClient interface {
+	GetSettings(options ...gitlab.RequestOptionFunc) (*gitlab.Settings, *gitlab.Response, error)
+}