+++ plugins/source/pagerduty/resources/services/users/user_contact_methods_fetch.go
@@ -10,7 +10,7 @@ import (
 	"github.com/cloudquery/plugin-sdk/schema"
 )
 
-func fetchUserContactMethods(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
+func fetchUserContactMethods(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
 	cqClient := meta.(*client.Client)
 	concreteParent := parent.Item.(pagerduty.User)