+++ scripts/table_diff/changes/changes_test.go
@@ -47,6 +47,26 @@ func Test_parseColumnChange(t *testing.T) {
 	}
 }
 
+func Test_parsePKChange(t *testing.T) {
+	type args struct {
+		line string
+	}
+	tests := []struct {
+		name      string
+		args      args
+		wantNames []string
+	}{
+		{name: "PK present", args: args{line: "The composite primary key for this table is (**org**, **id**, **hook_id**)."}, wantNames: []string{"org", "id", "hook_id"}},
+		{name: "no PK", args: args{}, wantNames: nil},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotNames := parsePKChange(tt.args.line)
+			require.Equal(t, tt.wantNames, gotNames)
+		})
+	}
+}
+
 func Test_getChanges(t *testing.T) {
 	tests := []struct {
 		name         string
@@ -175,6 +195,118 @@ func Test_getChanges(t *testing.T) {
 				},
 				{
 					Text:     "Table `gcp_resourcemanager_projects`: primary key constraint removed from column `_cq_id`",
+					Breaking: false,
+				},
+			},
+		},
+		{
+			name:         "Should mark newly added PK as breaking",
+			diffDataFile: "testdata/pr_5802_diff.txt",
+			wantChanges: []change{
+				{
+					Text:     "Table `aws_ses_configuration_sets`: column added with name `arn (PK)` and type `String`",
+					Breaking: true,
+				},
+				{
+					Text:     "Table `aws_ses_configuration_sets`: primary key constraint removed from column `account_id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `aws_ses_configuration_sets`: primary key constraint removed from column `name`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `aws_ses_configuration_sets`: primary key constraint removed from column `region`",
+					Breaking: false,
+				},
+			},
+		},
+		{
+			name:         "Should mark PK order change as breaking",
+			diffDataFile: "testdata/pr_6012_diff.txt",
+			wantChanges: []change{
+				{
+					Text:     "Table `github_external_groups`: column order changed for `updated_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_hook_deliveries`: primary key order changed from `org, id, hook_id` to `org, hook_id, id`",
+					Breaking: true,
+				},
+				{
+					Text:     "Table `github_hook_deliveries`: column order changed for `delivered_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_hook_deliveries`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_hooks`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_issues`: column added with name `state_reason` and type `String`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_organization_members`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_organizations`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_repositories`: column added with name `has_discussions` and type `Bool`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_repositories`: column order changed for `created_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_repositories`: column order changed for `pushed_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_repositories`: column order changed for `updated_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_members`: primary key order changed from `org, id, team_id` to `org, team_id, id`",
+					Breaking: true,
+				},
+				{
+					Text:     "Table `github_team_members`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_repositories`: primary key order changed from `org, id, team_id` to `org, team_id, id`",
+					Breaking: true,
+				},
+				{
+					Text:     "Table `github_team_repositories`: column added with name `has_discussions` and type `Bool`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_repositories`: column order changed for `created_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_repositories`: column order changed for `id`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_repositories`: column order changed for `pushed_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_team_repositories`: column order changed for `updated_at`",
+					Breaking: false,
+				},
+				{
+					Text:     "Table `github_workflows`: column order changed for `id`",
 					Breaking: false,
 				},
 			},