+++ plugins/destination/snowflake/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-snowflake-v1.1.0...plugins-destination-snowflake-v1.1.1) (2022-12-28)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.14.0 ([#6025](https://github.com/cloudquery/cloudquery/issues/6025)) ([35b2cfc](https://github.com/cloudquery/cloudquery/commit/35b2cfc7fc7bcdaceb7ee674e3a17f0f5673b366))
+
+## [1.1.0](https://github.com/cloudquery/cloudquery/compare/plugins-destination-snowflake-v1.0.7...plugins-destination-snowflake-v1.1.0) (2022-12-23)
+
+
+### Features
+
+* **destinations:** Migrate to managed batching SDK ([#5805](https://github.com/cloudquery/cloudquery/issues/5805)) ([2f130c1](https://github.com/cloudquery/cloudquery/commit/2f130c12c6e83ccd8a2d036ab5c47b55e2fb5280))
+
 ## [1.0.7](https://github.com/cloudquery/cloudquery/compare/plugins-destination-snowflake-v1.0.6...plugins-destination-snowflake-v1.0.7) (2022-12-20)